+++ src/tsp/solver/greedy.rs
@@ -59,14 +59,14 @@ impl GreedySolver {
         let target_distance = self.calculate_target_distance(); // Total available capacity / unvisited vertices
         let last_vertex = self.current_solution().get_last_vertex();
         for i in self.current_solution().unassigned_vertices() {   // Find vertex closest to the target distance
-            let difference = (self.instance().get_vertex(last_vertex as usize).get_weight(*i) as isize - target_distance as isize).abs() as usize; 
+            let difference = (self.instance().get_vertex(last_vertex).get_weight(*i) as isize - target_distance as isize).abs() as usize; 
             if difference < min_difference {
                 min_difference = difference;
                 best_vertex = *i;
             }
         }
 
-        (best_vertex, self.instance().get_vertex(last_vertex as usize).get_weight(best_vertex))
+        (best_vertex, self.instance().get_vertex(last_vertex).get_weight(best_vertex))
     }
 
     pub fn set_instance(&mut self, instance: &Rc<TSPInstance>) {
@@ -76,7 +76,7 @@ impl GreedySolver {
     pub fn solve_from_solution(&mut self, base_solution: Solution, next_vertex: u32) -> &Solution {
         self.current_solution = Some(base_solution);    // Initialize solution to the given one
         let next_driver = self.current_solution().get_smallest_driver();    // Find best driver
-        let distance = self.instance().get_vertex(self.current_solution().get_last_vertex() as usize).get_weight(next_vertex);  // Get distance between last vertex and chosen next vertex
+        let distance = self.instance().get_vertex(self.current_solution().get_last_vertex()).get_weight(next_vertex);  // Get distance between last vertex and chosen next vertex
         self.current_solution_mut().add_assignment(next_vertex, next_driver, distance); // Add the chosen vertex with the assigned driver to the solution
         self.solve_greedy();    // Solve the instance
         self.current_solution_mut().calculate_objective_value();    // Calculate the objective value of the solution
@@ -90,7 +90,7 @@ impl GreedySolver {
             self.current_solution_mut().add_assignment(best_vertex, next_driver, distance); // Add new vertex to the solution
         }
         let next_driver = self.current_solution().get_smallest_driver();    // Find best driver
-        let distance = self.instance().get_vertex(self.current_solution().get_last_vertex() as usize).get_weight(0);    // distance between the last assigned vertex and vertex 0
+        let distance = self.instance().get_vertex(self.current_solution().get_last_vertex()).get_weight(0);    // distance between the last assigned vertex and vertex 0
         self.current_solution_mut().add_assignment(0, next_driver, distance);   // Add final return to vertex 0
     }
 }
@@ -107,7 +107,7 @@ impl Solver for GreedySolver {
         println!("Val: {}", self.current_solution().objective_value());
         println!("Target: {}", self.instance().desired_travel_distance());
         println!("{:?}", self.current_solution().driver_distances());
-        
+
         logger.log_result(&self.current_solution());    // Log results
     }
